The great renaming
I really dislike it when developers do this, so it pains me to do it now,
but I have performed a mass renaming of classes.  FitsImage ended up being
the View in the MVC way of doing things, yet it shared the same naming
style as the model classes AstroImage and PythonImage.  This will be the
source of endless confusion to developers down the road.  So the renaming
went like this:

    FitsImage -> ImageView
    FitsImage{XYZ} -> ImageView{XYZ}
    PythonImage -> RGBImage

PythonImage also gets renamed to something much more akin to what it
actually is and/or implements.

So we have
   M: BaseImage, AstroImage, RGBImage
   V: ImageView{XYZ}
   C: Bindings, BindMap

I did this in the brand new 2.0 version so at least devs have a heads up
that things will not be backward compatible.

And I apologize in advance for any renaming and support issues this may
cause for you.