Added widget wrapper modules
I've added widget wrapper modules for gtk and qt.  These modules make it
possible to support simple plugins that do not have a complicated GUI with
a single non-toolkit specific plugin.

As a start the Ruler and Drawing plugins have been ported to use the new
modules, and have been removed from the qtw/plugins and gtkw/plugins.

Hopefully more plugins can be ported and this may make it easier to maintain
the plugins as well as to develop new ones that work for all supported widget
sets.