Fixed to reapply initial preferences if image changes
The initial application of preferences for zooming, cut levels, etc.
normally happens only when an image is visited for the first time.
This modification allows the viewer to track the image so that if the
image itself changes (e.g. size, data changed, etc) it will reapply
the initial preferences.

(reference viewer)
- updates to Pan to handle this
- updates to Mosaic to add experimental expansion feature so that
  mosaic image can expand if a new tile is added outside the region