Allow plugins and the GUI layout to be configured via a directory
This modification allows ginga plugins and the ginga layout to be
discovered via a directory.  In this first pass, the directory is
configured to be in the user's $HOME/.ginga, but this could be easily
changed in the future.

To write the set of standard plugins, invoke ginga with the command:

    $ ginga --write-plugin-conf

This will start ginga, but will also write the configuration for the
plugins to $HOME/.ginga/plugin_conf and a layout configuration to
$HOME/.ginga/layout.json

Upon subsequent starts, ginga will see that these files exist and will
use them to configure itself.