Allowed errors to print out from callbacks without loggers
Previously, when an error was thrown by a callback from an object without a
logger attribute, the error remained silent. I set it to print errors to
stderr when there is no available logger, instead. In doing so, I found an
error in gw/Desktop.py, where Desktop tries to log debug messages, but does
not have a logger. I fixed that by assigning it a logger in its __init__.