Added scroll with origin capability
This is inspired by the scroll zoom used in PanZoomCamera in VisPy.
Basically you zoom while keeping the coordinate under the cursor in the
same relative position in the window.

This can now be accessed by Shift+scroll in the empty mode and is the
default form for scroll in the "freepan" mode.