Improve plugin help by providing on-the-fly HTML conversion
This PR adds support for converting docstrings to HTML on the fly
so that the WBrowser plugin can display a more nicely formatted
help text.  In this initial PR, only "Cuts" plugin is configured to
show the help text in this way.

To use the plugin, the "docutils" package must be installed and the
WBrowser plugin must be loaded.