Add timed refresh option to the base Ginga viewer
This adds a convenience facility for doing timed refresh in a Ginga
viewer.  Using this it is possible to set up basic video type applications
where you update a numpy array with video frames from one thread (or a timer)
and the GUI thread refreshes as close to the desired rate as possible.

See the examples/gw/example1_video.py