Enhancement to Contents plugin to manage images in channels
This adds Load/Move/Copy/Remove buttons to the bottom of the Contents
UI.  Behavior of the Contents treeview is different.  First of all,
selecting a file does not display it.  Secondly, multiple files can be
selected, and using the standard Shift or Ctrl click the selection can
be modified.  An action is selected by clicking one of the buttons at
the bottom of the UI.  Move/Copy/Remove are all backed by pop-up dialogs
to confirm the operation.