Refactor shape implementation
Refactors shapes to have a more unified representation of points.
This allows more efficient sharing of code that operates of points.

Objects with (x, y) and (x1, y1, x2, y2) are refactored to store
the points in an array of points notation similar to polygon-based
shapes.  Properties are provided for backward-compatible access to
these attributes.