Addition of Quick Mode to Pick plugin
This adds a Quick Mode to the Pick plugin.  The addition works like this
(from the documentation):

After the area is moved, drawn or edited, Pick will perform one of three
actions:

1) In Quick Mode ON, with "From Peak" OFF, it will simply attempt to
   perform a calculation based on the coordinate under the crosshair in
   the center of the pick area.
2) In Quick Mode ON, with "From Peak" ON, it will perform a quick
   detection of peaks in the pick area and perform a calculation on the
   first one found, using the peak's coordinates.
3) In Quick Mode OFF, it will search the area for all peaks and
   evaluate the peaks based on the criteria in the "Settings" tab of the UI
   (see "The Settings Tab", below) and try to locate the best candidate
   matching the settings.

This mode also adds a plot to the set of plot tabs, called "Cuts".
It shows the cuts corresponding to the vertical and horizontal lines of
the crosshair, and updates in real time as the pick area is moved around
the image.

This commit includes documentation updates for the changes.