Add special "meta" mode and change the way modes are initiated
This adds a special "meta" mode, which is an intermediate mode used to
change modes.  In addition, the standard modes are all changed to be started
from meta mode.  Under the new scheme, you press Space to invoke meta mode,
then press the mode key that then invokes the mode you want.

The primary benefit of this modification is that it frees many key bindings
that can be used for other things, and allows more keys to be bound within
a mode.

The mode indicator has been updated to show a different color for the meta
mode and also to show an indicator for all mode types, not just "locked" (L)
and "softlock" (S), but also "held" (H) and "oneshot" (O).

This change affects the standard key bindings--documentation on quickref
and manual has been updated.  The old behavior can be retained by simply
installing the appropriate bindings.cfg file to $HOME/.ginga.