Efficiency improvement for Thumbs plugin
This changes the canvas drawing logic for the Thumbs plugin so that it
manages the thumbnails canvas much more efficiently, resulting in improved
usability and responsiveness.  The main improvement is that it calculates
which thumbnails should be visible based on the displayed area of the canvas
and adds/removes thumbs as necessary to only show that set.