Add a loader module
This adds a "first-level" loader module to ginga.  This module allows
opening a data item without knowing what it is first.  It works based
off of the mime type of the file and also the magic function, if available.

This loader works with both the reference viewer as well as when using
just the ginga widget.  This summarizes the difference:

Before:
  from ginga.AstroImage import AstroImage
  image = AstroImage(logger)
  image.load_file(some_path)

After:
  from ginga.util.loader import load_data
  image = load_data(path, logger=logger)

Although it is shorter, the main reason it is better is that in the
first instance we commit to creating a AstroImage, but if the file
references an RGB file this is not necessarily the right type.  In
the second case an AstroImage, a RGBImage or some other kind of image
might be returned.