Fixes for focus issued with MDI workspace mode
This fixes a issue with focus in both Qt and Gtk versions of MDI view
of a workspace.

For Qt, dragging a cursor into a channel viewer no longer forces a
page-switch event, which allows the `channel_follows_focus` option to
work correctly (not change the channel) if set to `False`.  The user
must either click in the window or title bar to change the channel.

For Gtk, this fixes an issue where if `enter_focus` is set to `False`,
then the widget would fail to ever get the keyboard focus.