Fix issue #677, for FutureWarning with numpy>=1.15.0
This fixes an issue for the error message issued by numpy >= 1.15.0:

FutureWarning: Using a non-tuple sequence for multidimensional indexing is
deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`. In the future this
will be interpreted as an array index, `arr[np.array(seq)]`, which will
result either in an error or a different result.