Refactor Zoom plugin
This small refactoring changes the way the Zoom plugin operates. Instead
of cutting a part of the data and updating a viewer, it shares the canvas
for the viewer that it is currently being used with, and simply sets the
pan position to the point of the cursor in the main viewer.

The advantages are:
- simplifies conceptual operation of the plugin
- makes the plugin more efficient
- allows the zoomed view to see items that are drawn on the canvas