Unify the coordinate API for canvas types
This adds alternative classes for many canvas types so that *all* types
use an API where points are specified as single parameters.  This will
be beneficial if we ever want to make a common change to coordinate
handling across all canvas objects.

To illustrate, before:

>>> canvas.add(Rectangle(x1, y1, x2, y2, ...))

after:

>>> canvas.add(RectangleP(p1, p2, ...))

Subclasses are used to keep API compatible versions of the old class
constructors available, and the new classes are not exposed as drawing
catalog members yet (although you can explictly import them and use them.
So there is no API change visible to any older code.