ViewerImageProxy: allow data operations on plotted mosaics (#812)
Add ViewerImageProxy class

Adds a ViewerImageProxy class, which is a kind of stand-in for an image (i.e. BaseImage-based) wrapper.  It looks at images plotted on the canvas and provides SOME methods to treat it as though it was one large single image.  This allows us to save large amounts of memory for mosaiced fields, because we don't need to allocate one huge numpy array to hold everything, but only plot each image at the proper location and orientation on the viewer's canvas--like a sparse data array.

Several plugins are modified so that they will work with this kind of mosaiced display, particularly `Pick`, `Cuts`, `Histogram`, `PixTable`.  A new plugin (called `Collage`) is added to build this kind of plotted mosaic.