Ginga pipeline (#845)
This is a new attempt at a generalized Pipeline plugin for Ginga.

A Pipeline class allows pipelines to be constructed of Stages. A Pipeline plugin is provided which offers a GUI for building and controlling a general image processing pipeline. A beginning set of stages are provided. I expect some interesting stages could be developed for quick look type applications like flat fielding, bias/dark/sky subtraction, binning, etc. based on astropy modules.

In addition to the full-image data pipelines, the standard pixel renderer is refactored to use a pipeline for generating the output to standard backends. This makes the rendering process much easier to understand and provides future opportunities for interesting modular hacks by inserting stages in the pipeline instead of having to carefully modify an algorithm embedded in the renderer itself.