Allow specification of stretch in list-spec of layout
This allows the specification of the stretch parameter in list-type
specifications of a layout.  Previously this was only possible in a
dict-type spec.  So you can now say, e.g.:
    ['hbox', dict(name='menu', stretch=0)],

instead of:
    dict(row=['hbox', dict(name='menu')],
           stretch=0),

This will simplify construction of layouts.

You can also specify 'border_width=' and 'spacing=' in parameter dicts
for vboxes and hboxes as well:

    ['vbox', dict(stretch=1, spacing=4, border_width=1),
      ...
    ],

Previous layouts should be unaffected by the change.