Pick plugin: account for coadds and multiple readouts in calculations
This PR adds the ability to account for image coadds and number of
detector readouts per coadd in Pick's calculations of background,
sky level and brightness.  The values for coadds and readouts are
read from the FITS header, if available.  The keywords associated
with these values can be set in the plugin settings.

This PR also moves the calculation of the star size into iqcalc, so
that the star size (size adjusting for pixel pitch of the detector)
is calculated for every pick candidate and not just the final one.