Add interactive setting of cut levels within Histogram plugin (#945)
Fix #942

You can set cut levels by clicking in the histogram plot:

* left click: set low cut
* middle click: reset (auto cut levels)
* right click: set high cut

In addition, you can dynamically adjust the gap between low and high cuts by scrolling the wheel in the plot (i.e. the "width" of the histogram plot curve).  This has the effect of increasing or decreasing the contrast within the image.  The amount that is changed for each wheel click is set by the plugin configuration file setting ``scroll_pct``.  The default is 10%.

* Adjust "plot by cuts" limits so that the cuts are visible