Modify Crosshair to contain the Cuts plot (#952)
This moves the "Cuts" plot and functionality from the Pick plugin to the Crosshair plugin.

Pick still has the "Quick Mode", which still works as the tool tips suggest, but simply does not activate a Cuts plot.

The Cuts plot uses the new ginga plots for good performance.  Documentation is updated in both plugins.