Change Contents so that "genthumb" setting of channel applies
This makes the Contents plugin in the reference viewer behave so
that if the 'genthumb' setting in the channel is set to False that
no entry is made in Contents. Additionally, if an individual image
has that setting made on it then no entry is added.

This makes Contents behave similarly to Thumbs for channel and image
settings.