Remove save_primary_header, inherit_primary_header attributes
This begins the removal of the save_primary_header and
inherit_primary_header attributes and kwargs from AstroImage.
The saving of the primary header is done when loading an image through
the io_fits module and it is merged when include_primary_header==True
in the AstroImage get_header() method.

The two kwargs issue a deprecation warning if used when creating an
AstroImage.

Additionally some other previously deprecated methods have been removed
from AstroImage.