Make pillow a required package, change piexif to exifread
Since we are planning to deprecate the mock backend, and pillow is used
in many places, as well as being a very common package, this adds pillow
as a requirement.  It also changes the optional piexif package to exifread,
which is a little less code verbose to use, a bit more popular and well
maintained.

This also restores a couple of lines needed in the OpenCv RGB file handler
to make auto-orientation of RGB images work correctly.