Fixed a longstanding issue for event handling on a ginga canvas
Fixed a longstanding issue where events registered on a canvas could
be masked by default key/cursor bindings (not associated with a mode).
This meant, for example, that only certain keystrokes could be
captured by an event handler registered on a ginga canvas, because any
keystrokes that had a default binding would take precedence.
Now such bindings are only executed if the event is not handled by any
active canvas bindings.