Remove the combination Viewer and Canvas (#973)
* Remove the combination Viewer and Canvas

The PR removes the combination Viewer-Canvas.  As a consequence, you cannot draw on a viewer, only a canvas and you cannot zoom a canvas, only a viewer.

This change has been phased in for a long time so that there shouldn't be too much legacy code that uses it. But it is not backward compatible as the combination classes have been eliminated.  Some of the older examples were still using this feature.  Users using the old classes will need to update their code.  For the most part, this is a rather straightforward change from ``ImageViewCanvas`` to ``CanvasView``.