Add default viewer profile (#1040)
This adds a feature called a "default viewer profile" to the ginga viewer widget.  Using this, the viewer can be optionally reset to some default settings between images.  This can be effectively used with the "remember" settings to have the viewer automatically recall the exact way an image was viewed previously (when viewed again), but have the settings restored to some reasonable default when viewing a new image.

This also separates out the color distribution algorithm and contrast settings as "rememberable" and "default settable" items separate from the color map settings.

The way the settings are applied when changing to an image is as follows:

1) the viewer is reset to the default viewer profile, if any
2) any remembered items from the image profile are applied, if any
3) any auto adjustments cuts/zoom/center are applied, if they were not overridden by a remembered setting

* Fix application of viewer profile and image profile
* Update documentation on Preferences
* Update figure sizes in documentation
* Update WhatsNew