Update event handling at the lower ginga viewer widget level
The event handling API at the ginga viewer lower level is changed
so that event objects are passed for all event callbacks at this level.
This includes the following callbacks: map, focus, enter, leave,
key-press, key-release, button-press, button-release, motion, scroll,
swipe, pinch, pan

Note that this level of API use is recommended only for users adding
a viewer widget to their own viewer.  Users of the reference viewer
should be using the callbacks in ginga.Bindings, which have not changed.

The reason for the change is to allow more flexible changes to the
event system by changing the event in the future instead of adding
fixed position arguments to the callback API.